@rynowak 2019-04-12
> +3% gain on mvc plaintext

This is super dope!