@rynowak 2019-04-15
Exceptions look good!. Yeah it would be awesome if you want to do this for the other invokers.