@benaadams 2019-04-19
>  Yeah it would be awesome if you want to do this for the other invokers.

ControllerInvoker https://github.com/aspnet/AspNetCore/pull/9533